Settings: Split settings to game and compatibility settings
